"Update color property of .SearchBar input selector"
In the src/components/SearchBar/SearchBar.css file, the color property of the .SearchBar input selector has been updated. The color has been changed from #010c3f to green. This change also applies to the .SearchBar input selector within the media query for max-width: 600px.

The motivation for this change was to improve the visibility and user experience of the search bar. The previous color, #010c3f, was found to be too dark and made the input text difficult to read. By changing the color to green, the input text is now more visible and easier to read, especially on smaller screens where the max-width is 600px.