"Add savePlaylist function and update onClick handler in Playlist.js"
This commit introduces a new function, savePlaylist, to the Playlist component in src/components/Playlist/Playlist.js. The savePlaylist function maps over the playlistTracks in the state to create an array of track URIs, which is then passed to the savePlaylist function from the Spotify module along with the playlist name from the state.

In addition, the onClick handler for the 'SAVE TO SPOTIFY' button has been updated to call this new savePlaylist function.

The motivation for these changes was to enhance the functionality of the Playlist component by allowing users to save their playlists to Spotify. This differs from the previous implementation where the 'SAVE TO SPOTIFY' button did not have any functionality attached to it.

These changes were inspired by several tasks in the build_module_backend.services.TaskCreation module.