"Add CSS class to SearchBar input element"
In the file src/components/SearchBar/SearchBar.js, a CSS class 'search-bar-text' has been added to the input element in the render method. This class is defined in the CSS file and is used to change the color of the search bar text.

The motivation for this change was to improve the visual appearance and user experience of the search bar. Previously, the search bar text color was not customizable, and this change allows for more flexibility in styling. This differs from the previous implementation where the text color was static and could not be changed through CSS.