"Update onSearch prop in SearchBar component"
In the file src/components/App/App.js, the onSearch prop of the SearchBar component has been updated from this.openAiSearch to this.search. This change was made to reflect the modifications in the SearchBar component.

The motivation for this change was to ensure that the onSearch prop correctly references the updated search function in the SearchBar component. Previously, it was referencing the old openAiSearch function, which has now been replaced with the search function. This change ensures that the onSearch prop is up-to-date with the latest changes in the SearchBar component.