"Add className attribute to input element in SearchBar.js"
This commit adds a className attribute to the input element in the render method of the SearchBar.js file. The className is set to 'search-bar-text', which should be the CSS class or id associated with the search bar text.

The motivation for this change was to provide a way to style the search bar text using CSS. This differs from the previous implementation where there was no className attribute set for the input element, limiting the ability to apply specific styles to the search bar text.