"Update Playlist.js to include SearchBar component"
In this commit, the SearchBar component has been imported from the correct path and added to the render method of the Playlist component. This change was motivated by tasks that required the SearchBar component to be accessible within the Playlist component. The previous implementation did not include the SearchBar component in the Playlist component's render method. This update ensures that the SearchBar component is correctly rendered within the Playlist component.