Include genre in Spotify track object
This commit adds genre information to the Spotify track objects returned by the map function in src/util/Spotify.js. The motivation for this change is to provide more detailed track information to the end user, enhancing the overall user experience by allowing them to see the genre of each track.

Previously, the returned track object did not contain genre information. Now, the genre is retrieved from the track.album.genres property. If the genre information is an array, it is joined into a single string. This ensures that the genre data is presented in a consistent and readable format.