Add genre property to track object
This commit introduces a new 'genre' property to the track object within the map function in src/util/Spotify.js. The motivation for this change is to enhance the track information by including genre details, which can be useful for various features such as creating more refined playlists or displaying more detailed track information to the user.

Previously, the track object did not contain genre information. Now, the genre is extracted from the artist's genres array. If the artist's genres array is not available or is empty, the genre is set to 'Unknown'. This ensures that every track object has a genre property, improving the consistency of the track data presented to the users.